pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sand-atlas"
version = "1.3.0"
authors = [
    {name = "Benjy Marks", email = "benjy.marks@sydney.edu.au"},
    {name = "Ilija Vego", email = "ilija.vego@sydney.edu.au"}
]
description = "A python package for processing 3d volumetric data to build meshes for the Sand Atlas data repository"
dependencies = [
    "black",
    "flake8",
    "numpy",
    "pandas",
    "scikit-image",
    "sphinx",
    "nbsphinx",
    "tifffile",
    "pynrrd",
    "h5py",
    "spam",
    "tqdm",
    "colorama",
]

requires-python = ">=3.8"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"
]
keywords = ["3d", "volumetric", "meshes", "data", "granular", "sand", "particles"]

[project.optional-dependencies]
neuroglancer  = [ "cloud-volume" ]
dev = ["ipython", "jupyter", "pytest", "pytest-cov"]

[project.scripts]
sand_atlas_process = "sand_atlas.pipeline:full_analysis_script"
sand_atlas_properties = "sand_atlas.pipeline:properties_script"
vdb_to_npy = "sand_atlas.pipeline:vdb_to_npy"
sand_atlas_clean_labels = "sand_atlas.pipeline:clean_labels_script"
sand_atlas_preflight = "sand_atlas.preflight:preflight_script"

[project.urls]
Homepage = "https://sand-atlas.scigem.com"
Documentation = "https://github.com/scigem/sand-atlas-python/docs/index.html"
Repository = "https://github.com/scigem/sand-atlas-python.git"

[tool.black]
line-length = 120
target-version = ['py312']

[tool.flake8]
max-line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers"
markers = [
    "unit: fast isolated tests for library logic",
    "feature: user-facing behavior tests for scripts and workflows",
    "integration: multi-step pipeline integration tests",
    "external: tests that require external tools such as Blender or FFmpeg",
    "network: tests that exercise network-facing code paths",
    "slow: long-running tests excluded from the default developer loop",
]

[tool.coverage.run]
source = ["sand_atlas"]
omit = [
    "sand_atlas/blender_scripts/*",
    "sand_atlas/matlab_scripts/*",
]

[tool.coverage.report]
skip_empty = true
show_missing = true
fail_under = 75
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
